// src/types/Controller/Booking.ts
interface Reservation {
  id: number;
  'reference-id': string | null;
  type: string;
  arrival: string;
  departure: string;
  'created-at': string;
  modifiedAt: string;
  apartment: { id: number; name: string };
  channel: { id: number; name: string };
  'guest-name': string | null;
  firstname: string;
  lastname: string;
  email: string | null;
  phone: string | null;
  adults: number;
  children: number;
  'check-in': string;
  'check-out': string;
  notice: string;
  'assistant-notice': string;
  price: number;
  'price-details': string;
  'city-tax': number | null;
  'price-paid': string;
  'commission-included': number | null;
  prepayment: number | null;
  'prepayment-paid': string;
  deposit: number | null;
  'deposit-paid': string;
  language: string;
  'guest-app-url': string;
  'is-blocked-booking': boolean;
  guestId: number | null;
  related: { id: number; name: string }[];
}
export default Reservation;